pyproject.toml
[project]
name = "clasificacion-etiquetas-borrador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "langgraph-cli[inmem]>=0.4.12",
]
[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
